requirements.txt
python-dotenv==0.15.0
python-rapidjson==1.6
Werkzeug==2.0.2
loguru==0.5.3
pyjwt==2.0.1
requests==2.25.1
psycopg2-binary==2.8.6
cerberus==1.3.4
boto3==1.21.21
botocore==1.24.21
croniter==1.0.12
pandas==1.3.5
kafka-python==2.0.2
smart_open[s3]==5.1.0
Flask==2.0.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
flask-restx==0.5.1
aioboto3==10.0.0
docker==6.0.0
coverage==6.4.4
pytest==7.1.3
typer==0.6.1
tabulate==0.8.10
